requirements.txt
google-generativeai
python-dotenv
PyMuPDF
requests
streamlit
pandas
plotly
scikit-learn
